[Improve] Move SimpleChecklistModel to domain and remove model reference from HomeState because uuid was unnecessary on ui layer (just used on ViewModel to delete checklist)
