+++ app/src/main/java/wottrich/github/io/smartchecklist/domain/mapper/SimpleChecklistModelMapper.kt
@@ -1,7 +1,7 @@
 package wottrich.github.io.smartchecklist.domain.mapper
 
 import wottrich.github.io.smartchecklist.datasource.data.model.Checklist
-import wottrich.github.io.smartchecklist.presentation.ui.model.SimpleChecklistModel
+import wottrich.github.io.smartchecklist.domain.model.SimpleChecklistModel
 
 class SimpleChecklistModelMapper {
     fun mapToSimpleChecklistModel(checklist: Checklist) =
