+++ app/src/main/java/wottrich/github/io/smartchecklist/domain/usecase/ObserveSimpleSelectedChecklistModelUseCase.kt
@@ -1,14 +1,14 @@
 package wottrich.github.io.smartchecklist.domain.usecase
 
-import wottrich.github.io.smartchecklist.coroutines.FlowableUseCase
-import wottrich.github.io.smartchecklist.coroutines.UseCase
-import wottrich.github.io.smartchecklist.coroutines.base.Result
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
-import wottrich.github.io.smartchecklist.domain.mapper.SimpleChecklistModelMapper
-import wottrich.github.io.smartchecklist.presentation.ui.model.SimpleChecklistModel
 import wottrich.github.io.smartchecklist.checklist.data.repository.ChecklistRepository
+import wottrich.github.io.smartchecklist.coroutines.FlowableUseCase
+import wottrich.github.io.smartchecklist.coroutines.UseCase
+import wottrich.github.io.smartchecklist.coroutines.base.Result
+import wottrich.github.io.smartchecklist.domain.mapper.SimpleChecklistModelMapper
+import wottrich.github.io.smartchecklist.domain.model.SimpleChecklistModel
 
 class ObserveSimpleSelectedChecklistModelUseCase(
     private val checklistRepository: ChecklistRepository,
