+++ app/src/main/java/wottrich/github/io/smartchecklist/presentation/state/HomeState.kt
@@ -2,11 +2,10 @@ package wottrich.github.io.smartchecklist.presentation.state
 
 import wottrich.github.io.smartchecklist.presentation.state.HomeUiState.Loading
 import wottrich.github.io.smartchecklist.presentation.state.HomeUiState.Overview
-import wottrich.github.io.smartchecklist.presentation.ui.model.SimpleChecklistModel
 
 data class HomeState(
     val homeUiState: HomeUiState,
-    val checklist: SimpleChecklistModel?
+    val checklistName: String?
 ) {
     val isEditUiState: Boolean
         get() = homeUiState is Overview && homeUiState.isEditing
@@ -18,7 +17,7 @@ data class HomeState(
     companion object {
         val Initial = HomeState(
             homeUiState = Loading,
-            checklist = null
+            checklistName = null
         )
     }
 }
